Refactor page selectors
